refactor: Separating components
