+++ client/src/components/Category/index.js
@@ -1,0 +1,4 @@
+export { default as Header } from './Header';
+export { default as Filters } from './Filters';
+export { default as FiltersContainer } from './FiltersContainer';
+export { default as ResultProducts } from './ResultProducts';
