+++ client/src/components/Products/Description.jsx
@@ -1,0 +1,9 @@
+import { Text } from '@mantine/core';
+
+const Description = ({ children }) => (
+  <Text fz="1.6rem" fw="500" lh="3.2rem" mt="1.5rem">
+    {children}
+  </Text>
+);
+
+export default Description;
