+++ client/src/components/Products/Title.jsx
@@ -1,0 +1,9 @@
+import { Text } from '@mantine/core';
+
+const Title = ({ children }) => (
+  <Text fz="2.8rem" fw="600">
+    {children}
+  </Text>
+);
+
+export default Title;
