+++ client/src/components/Products/index.js
@@ -1,0 +1,5 @@
+export { default as Title } from './Title';
+export { default as Info } from './Info';
+export { default as Description } from './Description';
+export { default as CartButton } from './CartButton';
+export { default as WishListButton } from './WishListButton';
